requirements.txt
amqp==5.3.1
asgiref==3.8.0
beautifulsoup4==4.12.2
billiard==4.2.4
black==23.3.0
boto3==1.34.54
botocore==1.34.162
brotli==1.2.0
celery==5.3.6
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
channels==4.0.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
cron_descriptor==2.0.6
cryptography==42.0.5
cssselect==1.3.0
cssselect2==0.8.0
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.4.0
dj-database-url==2.1.0
Django==4.2.11
django-allauth==0.57.0
django-appconf==1.2.0
django-auditlog==2.3.0
django-axes==6.0.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-csp==4.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-fsm==2.8.1
django-fsm-log==3.1.0
django-health-check==3.17.0
django-htmx==1.17.3
django-import-export==3.3.0
django-js-asset==3.1.2
django-prometheus==2.3.1
django-ratelimit==4.0.0
django-redis==5.3.0
django-storages==1.14.2
django-taggit==4.0.0
django-timezone-field==7.2.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et_xmlfile==2.0.0
filelock==3.20.3
flake8==6.0.0
fonttools==4.61.1
goose3==3.1.14
gunicorn==21.2.0
html5lib==1.1
identify==2.6.16
idna==3.11
inflection==0.5.1
isort==5.12.0
jmespath==1.1.0
kombu==5.6.2
langdetect==1.0.9
language-tool-python==2.7.1
lxml==4.9.3
MarkupPy==1.18
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
nodeenv==1.10.0
numpy==1.26.4
oauthlib==3.3.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==26.0
pandas==2.2.1
pathspec==1.0.3
pdfkit==1.0.0
Pillow==10.1.0
platformdirs==4.5.1
pre-commit==3.3.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2-binary==2.9.9
pyahocorasick==2.3.0
pyClamd==0.4.0
pycodestyle==2.10.0
pycparser==3.0
pydyf==0.12.1
pyflakes==3.0.1
PyJWT==2.10.1
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
readability-lxml==0.8.1
redis==5.0.1
reportlab==4.0.9
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.10.4
sentry-sdk==1.40.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
tablib==3.5.0
tinycss2==1.5.1
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==1.26.20
vine==5.1.0
virtualenv==20.36.1
wcwidth==0.3.5
weasyprint==61.0
webencodings==0.5.1
websockets==12.0
whitenoise==6.6.0
xlrd==2.0.2
xlwt==1.3.0
zopfli==0.4.0

stripe-14.2.0
